+++ src/main.rs
@@ -1,8 +1,5 @@
 use parking_lot::Mutex;
-use std::{
-    collections::HashMap,
-    sync::{Arc, atomic::AtomicBool},
-};
+use std::sync::Arc;
 
 use anyhow::{Context, anyhow, bail};
 use chrono::{DateTime, Local};
@@ -10,52 +7,26 @@ use teloxide::{
     dispatching::dialogue::{GetChatId, InMemStorage},
     payloads::SetMessageReactionSetters,
     prelude::*,
-    types::{
-        InlineKeyboardButton, InlineKeyboardMarkup, InputFile, InputPollOption, Location,
-        MessageId, ReactionType, UpdateKind,
-    },
+    types::{InlineKeyboardButton, InlineKeyboardMarkup, InputFile, InputPollOption, ReactionType},
 };
 
 use crate::{
-    polls::{MainQuestion, QuestionaireQuestion},
+    polls::MainQuestion,
     questionaire::QuestionaireFrogName,
-    weather::{BotWeatherExt, WeatherStats},
+    state::State,
+    utils::{DialogueState, LastLocation, MessageClassification, Mode, SentMessage, TimedLocation},
+    weather::WeatherStats,
 };
 
 mod counting;
 mod polls;
 mod questionaire;
 mod reports;
+mod state;
+mod utils;
 mod weather;
 
 const TOKEN: &'static str = include_str!("../token.txt").trim_ascii();
-type DialogueState = Dialogue<State, InMemStorage<State>>;
-
-trait PollExt {
-    fn selected(&self) -> &str;
-    fn selected_index(&self) -> isize;
-}
-
-impl PollExt for Poll {
-    fn selected(&self) -> &str {
-        self.options
-            .iter()
-            .filter(|o| o.voter_count > 0)
-            .map(|o| o.text.as_str())
-            .next()
-            .unwrap_or_default()
-    }
-
-    fn selected_index(&self) -> isize {
-        self.options
-            .iter()
-            .enumerate()
-            .filter(|(_, o)| o.voter_count > 0)
-            .map(|(i, _)| i as isize)
-            .next()
-            .unwrap_or(-1)
-    }
-}
 
 #[derive(Debug, Clone, Copy, serde::Deserialize, serde::Serialize)]
 pub enum Sex {
@@ -148,107 +119,8 @@ impl CompleteWalk {
             location,
             source,
             time: Local::now(),
-            gps_location: if_is_relevant(gps_location),
+            gps_location: gps_location.as_location(),
         });
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Mode(Arc<AtomicBool>);
-
-#[derive(Debug, Clone, Copy, Default, serde::Serialize, serde::Deserialize)]
-pub enum MessageClassification {
-    #[default]
-    None,
-    Weather,
-    Frog(usize),
-    DeadFrog(usize),
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct SentMessage {
-    context: Arc<parking_lot::Mutex<HashMap<MessageId, MessageClassification>>>,
-    history: Arc<parking_lot::Mutex<Vec<MessageId>>>,
-}
-
-impl SentMessage {
-    fn clear_history(&self) {
-        self.history.lock().clear();
-    }
-    fn clear(&self) {
-        self.context.lock().clear();
-        self.history.lock().clear();
-    }
-
-    fn add_weather(&self, id: MessageId) {
-        self.add_to_history(id);
-        self.context
-            .lock()
-            .insert(id, MessageClassification::Weather);
-    }
-
-    fn add_frog(&self, id: MessageId, index: usize) {
-        self.add_to_history(id);
-        self.context
-            .lock()
-            .insert(id, MessageClassification::Frog(index));
-    }
-
-    fn add_dead_frog(&self, id: MessageId, index: usize) {
-        self.add_to_history(id);
-        self.context
-            .lock()
-            .insert(id, MessageClassification::DeadFrog(index));
-    }
-
-    fn get(&self, id: MessageId) -> Option<MessageClassification> {
-        self.context.lock().get(&id).copied()
-    }
-
-    fn add_to_history(&self, id: MessageId) {
-        self.history.lock().push(id);
-    }
-
-    async fn go_back(
-        &self,
-        bot: Bot,
-        dialoge: Dialogue<State, InMemStorage<State>>,
-    ) -> anyhow::Result<()> {
-        let Some(id) = self.history.lock().pop() else {
-            eprintln!("No going back possible!");
-            return Ok(());
-        };
-        let r = bot.delete_message(dialoge.chat_id(), id).await?;
-        dbg!(r);
-        let previous = dialoge.get_or_default().await?.go_back();
-        dialoge.update(previous).await?;
-        Ok(())
-    }
-}
-
-impl Mode {
-    pub fn change_to_debug(&self) {
-        self.0.store(true, std::sync::atomic::Ordering::SeqCst);
-    }
-    pub fn change_to_release(&self) {
-        self.0.store(false, std::sync::atomic::Ordering::SeqCst);
-    }
-    pub fn as_path(&self) -> &'static str {
-        if self.is_debug() {
-            "debug-walks"
-        } else {
-            "walks"
-        }
-    }
-
-    pub fn create_debug() -> Self {
-        Self(Arc::new(AtomicBool::new(true)))
-    }
-    pub fn create_release() -> Self {
-        Self(Arc::new(AtomicBool::new(false)))
-    }
-    pub fn is_debug(&self) -> bool {
-        self.0.load(std::sync::atomic::Ordering::SeqCst)
     }
 }
 
@@ -296,430 +168,9 @@ impl PartialFrog {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-pub enum State {
-    #[default]
-    Start,
-    WalkStarted {
-        walk: CompleteWalk,
-    },
-    WaitForModeChange {
-        walk: CompleteWalk,
-    },
-    QuestionaireFrogName {
-        walk: CompleteWalk,
-        questionaire: questionaire::QuestionaireFrogName,
-    },
-    QuestionaireSex {
-        walk: CompleteWalk,
-        questionaire: questionaire::QuestionaireSex,
-    },
-    DeadFrog {
-        walk: CompleteWalk,
-    },
-    DeadFrogName {
-        walk: CompleteWalk,
-        name: Option<String>,
-    },
-    FrogIdentified {
-        frog: PartialFrog,
-        walk: CompleteWalk,
-    },
-    EnterTemperature {
-        is_start: bool,
-        prev_state: Box<State>,
-    },
-    ChangePercipation {
-        prev_state: Box<State>,
-    },
-}
-
-impl State {
-    fn as_walk(&self) -> Option<CompleteWalk> {
-        match self {
-            State::Start => None,
-            State::WalkStarted { walk } | State::WaitForModeChange { walk } => Some(walk.clone()),
-            State::QuestionaireFrogName { walk, .. } => Some(walk.clone()),
-            State::QuestionaireSex { walk, .. } => Some(walk.clone()),
-            State::DeadFrog { walk } => Some(walk.clone()),
-            State::DeadFrogName { walk, .. } => Some(walk.clone()),
-            State::FrogIdentified { walk, .. } => Some(walk.clone()),
-            State::ChangePercipation { .. } | State::EnterTemperature { .. } => todo!(),
-        }
-    }
-
-    fn as_walk_mut(&mut self) -> Option<&mut CompleteWalk> {
-        match self {
-            State::Start => None,
-            State::WalkStarted { walk } | State::WaitForModeChange { walk } => Some(walk),
-            State::QuestionaireFrogName { walk, .. } => Some(walk),
-            State::QuestionaireSex { walk, .. } => Some(walk),
-            State::DeadFrog { walk } => Some(walk),
-            State::DeadFrogName { walk, .. } => Some(walk),
-            State::FrogIdentified { walk, .. } => Some(walk),
-            State::ChangePercipation { prev_state }
-            | State::EnterTemperature { prev_state, .. } => prev_state.as_walk_mut(),
-        }
-    }
-
-    async fn enter_temperature(
-        bot: Bot,
-        dialoge: DialogueState,
-        (is_start, prev_state): (bool, Box<State>),
-        message: Message,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        let temp = message
-            .text()
-            .ok_or(anyhow!("There should be a text message???"))?;
-        let temp: f64 = temp.parse()?;
-        let mut state = *prev_state;
-        let weather = &mut state
-            .as_walk_mut()
-            .expect("Should be set at this point?")
-            .weather;
-        if is_start {
-            weather.temperature_start = temp;
-        } else {
-            weather.temperature_end = Some(temp);
-        }
-        let m = bot.send_weather_stats(dialoge.chat_id(), *weather).await?;
-        sent.add_weather(m.id);
-        dialoge.update(state).await?;
-        Ok(())
-    }
-
-    async fn start(bot: Bot, dialoge: DialogueState, sent: SentMessage) -> anyhow::Result<()> {
-        sent.clear();
-        let walk = CompleteWalk::start()
-            .await
-            .context("Creating walk for new walk created by user")?;
-        bot.send_weather_stats(dialoge.chat_id(), walk.weather)
-            .await
-            .context("Sending the weather via tg to user")?;
-        dialoge.update(State::WalkStarted { walk }).await?;
-        Ok(())
-    }
-
-    async fn change_percipation(
-        bot: Bot,
-        prev_state: Box<State>,
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        use weather::Percipation::*;
-        let percipation = match poll.selected_index() {
-            0 => None,
-            1 => Fog,
-            2 => Drizzle,
-            3 => ModerateRain,
-            4 => StrongRain,
-            5 => Graupel,
-            6 => Snow,
-            -1 => bail!("TODO, no unselecting allowed!"),
-            _ => unreachable!(),
-        };
-        let mut state = *prev_state;
-        let weather = &mut state.as_walk_mut().expect("Should be unreachable").weather;
-        weather.percipation = percipation;
-        let w = bot
-            .send_weather_stats(dialoge.chat_id(), *weather)
-            .await?
-            .id;
-        sent.add_weather(w);
-        dialoge.update(state).await?;
-        Ok(())
-    }
-
-    async fn poll_answer_walk_started(
-        bot: Bot,
-        last_location: LastLocation,
-        walk: CompleteWalk,
-        dialoge: DialogueState,
-        poll: Poll,
-        mode: Mode,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        if poll.selected_index() < 0 {
-            sent.go_back(bot, dialoge).await?;
-            return Ok(());
-        }
-        match poll.selected() {
-            "End" => {
-                maybe_end_walk(bot, walk, dialoge, mode, sent).await?;
-            }
-            name @ ("Erdkröte" | "Grasfrosch" | "Teichmolch" | "Bergmolch" | "Kammmolch") => {
-                let last_message_id = MainQuestion::AskForSex(name.into())
-                    .ask(bot, dialoge.chat_id())
-                    .await?;
-                sent.add_frog(last_message_id, walk.frogs.len());
-                dialoge
-                    .update(State::FrogIdentified {
-                        frog: PartialFrog {
-                            name: name.into(),
-                            gps_location: if_is_relevant(last_location),
-                            ..Default::default()
-                        },
-                        walk,
-                    })
-                    .await?;
-            }
-            "Found Something" => {
-                let last_message_id = QuestionaireQuestion::IsItAFrogToadOrMolch
-                    .ask(bot.clone(), dialoge.chat_id())
-                    .await?;
-                sent.add_frog(last_message_id, walk.frogs.len());
-                dialoge
-                    .update(State::QuestionaireFrogName {
-                        walk,
-                        questionaire: QuestionaireFrogName::new(),
-                    })
-                    .await?;
-            }
-            "Dead Frog :(" => {
-                let id = MainQuestion::FoundDeadFrog
-                    .ask(bot, dialoge.chat_id())
-                    .await?;
-                sent.add_dead_frog(id, walk.dead_frogs.len());
-                dialoge.update(State::DeadFrog { walk }).await?;
-            }
-            _ => bail!("TODO"),
-        }
-        Ok(())
-    }
-
-    async fn dead_frog_answered(
-        bot: Bot,
-        walk: CompleteWalk,
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        if poll.selected_index() < 0 {
-            sent.go_back(bot, dialoge).await?;
-            return Ok(());
-        }
-        let name = match poll.selected() {
-            "No" => None,
-            name => Some(name.to_string()),
-        };
-        let id = MainQuestion::WhereAreYou
-            .ask(bot, dialoge.chat_id())
-            .await?;
-        sent.add_dead_frog(id, walk.dead_frogs.len());
-        dialoge.update(State::DeadFrogName { walk, name }).await?;
-        Ok(())
-    }
-
-    async fn dead_frog_location_answered(
-        bot: Bot,
-        (mut walk, name): (CompleteWalk, Option<String>),
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        let location = poll.selected_index();
-        if location < 0 {
-            sent.go_back(bot, dialoge).await?;
-            return Ok(());
-        }
-        let location = location as usize;
-        walk.dead_frogs.push(DeadFrog {
-            name,
-            sex: None,
-            location,
-            time: Local::now(),
-        });
-        dialoge.update(State::WalkStarted { walk }).await?;
-        sent.clear_history();
-        sent.add_to_history(
-            MainQuestion::FoundSomething
-                .ask(bot, dialoge.chat_id())
-                .await?,
-        );
-        Ok(())
-    }
-
-    async fn frog_identified(
-        bot: Bot,
-        (mut frog, walk): (PartialFrog, CompleteWalk),
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        if poll.selected_index() < 0
-        // && let Some(q) = MainQuestion::find_original(&poll.question)
-        {
-            sent.go_back(bot, dialoge).await?;
-            // match q {
-            //     MainQuestion::FoundSomething => {
-            //         dialoge.update(State::WalkStarted { walk }).await?;
-            //     }
-            //     MainQuestion::FoundDeadFrog => unreachable!("I think this is unreachable"),
-            //     MainQuestion::WhereIsFrogHeaded => frog.towards = None,
-            //     MainQuestion::WhereAreYou => frog.location = None,
-            //     MainQuestion::AskForSex(_) => frog.sex = None,
-            // }
-            // bot.delete_message(dialoge.chat_id(), last_message_id)
-            //     .await?;
-            return Ok(());
-        }
-        if frog.location.is_some() {
-            State::frog_identified_sex_location(bot, (frog, walk), dialoge, poll, sent).await?;
-        } else if frog.sex.is_some() {
-            State::frog_identified_sex(bot, (frog, walk), dialoge, poll, sent).await?;
-        } else {
-            let sex = match poll.selected() {
-                "Male" => Sex::Male,
-                "Female" => Sex::Female,
-                "Unknown" => Sex::Unknown,
-                "Use Questionaire" => {
-                    let last_message_id =
-                        questionaire::start_sex(bot, dialoge.chat_id(), &frog.name).await?;
-                    sent.add_frog(last_message_id, walk.frogs.len());
-
-                    dialoge
-                        .update(State::QuestionaireSex {
-                            walk,
-                            questionaire: questionaire::QuestionaireSex::new(frog.clone()),
-                        })
-                        .await?;
-                    return Ok(());
-                }
-                _ => unreachable!(),
-            };
-            frog.sex = Some(sex);
-            let last_message_id = MainQuestion::WhereAreYou
-                .ask(bot, dialoge.chat_id())
-                .await?;
-            sent.add_frog(last_message_id, walk.frogs.len());
-            dialoge.update(State::FrogIdentified { frog, walk }).await?;
-        }
-        Ok(())
-    }
-
-    async fn frog_identified_sex(
-        bot: Bot,
-        (mut frog, walk): (PartialFrog, CompleteWalk),
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        let location = poll.selected_index();
-        if location < 0 {
-            bail!("We do not really handle you unselecting something. sorry.");
-        }
-        let location = location as usize;
-        frog.location = Some(location);
-        let last_message_id = MainQuestion::WhereIsFrogHeaded
-            .ask(bot, dialoge.chat_id())
-            .await?;
-        sent.add_frog(last_message_id, walk.frogs.len());
-        dialoge.update(State::FrogIdentified { frog, walk }).await?;
-        Ok(())
-    }
-
-    async fn frog_identified_sex_location(
-        bot: Bot,
-        (mut frog, mut walk): (PartialFrog, CompleteWalk),
-        dialoge: DialogueState,
-        poll: Poll,
-        sent: SentMessage,
-    ) -> anyhow::Result<()> {
-        let towards = poll.selected_index() == 0;
-        if poll.selected_index() < 0 {
-            bail!("We do not really handle you unselecting something. sorry.");
-        }
-        frog.towards = Some(towards);
-        let frog = frog.build()?;
-        let id = bot
-            .send_message(dialoge.chat_id(), format!("Found {}", frog.to_message()))
-            .reply_markup(InlineKeyboardMarkup::new([
-                [InlineKeyboardButton::callback("Repeat", "found:repeat")],
-                [InlineKeyboardButton::callback("Find", "found:next")],
-                [InlineKeyboardButton::callback("End", "found:end")],
-            ]))
-            .await?
-            .id;
-        sent.add_frog(id, walk.frogs.len());
-        walk.frogs.push(frog);
-        walk.repeats = 1;
-        dialoge.update(State::WalkStarted { walk }).await?;
-        Ok(())
-    }
-
-    fn go_back(&self) -> Self {
-        match self {
-            State::WalkStarted { .. } | State::Start => self.clone(),
-            State::DeadFrog { walk } | State::WaitForModeChange { walk } => {
-                Self::WalkStarted { walk: walk.clone() }
-            }
-            State::QuestionaireFrogName { walk, questionaire } => {
-                if let Some(questionaire) = questionaire.go_back() {
-                    Self::QuestionaireFrogName {
-                        walk: walk.clone(),
-                        questionaire,
-                    }
-                } else {
-                    Self::WalkStarted { walk: walk.clone() }
-                }
-            }
-            State::QuestionaireSex { walk, questionaire } => {
-                if let Some(questionaire) = questionaire.go_back() {
-                    Self::QuestionaireSex {
-                        walk: walk.clone(),
-                        questionaire,
-                    }
-                } else {
-                    Self::FrogIdentified {
-                        frog: questionaire.frog.clone(),
-                        walk: walk.clone(),
-                    }
-                }
-            }
-            State::DeadFrogName { walk, name } => {
-                if name.is_some() {
-                    Self::DeadFrogName {
-                        walk: walk.clone(),
-                        name: None,
-                    }
-                } else {
-                    Self::WalkStarted { walk: walk.clone() }
-                }
-            }
-            State::FrogIdentified { frog, walk } => {
-                if let Some(frog) = frog.go_back() {
-                    Self::FrogIdentified {
-                        frog,
-                        walk: walk.clone(),
-                    }
-                } else {
-                    // TODO: Check that using questionaire does not break
-                    // anything here!
-                    Self::WalkStarted { walk: walk.clone() }
-                }
-            }
-            State::ChangePercipation { prev_state }
-            | State::EnterTemperature { prev_state, .. } => *prev_state.clone(),
-        }
-    }
-}
-
-fn if_is_relevant(last_location: LastLocation) -> Option<TimedLocation> {
-    let last_location = last_location.lock();
-    if last_location.latitude.is_nan() || last_location.longitude.is_nan() {
-        None
-    } else if (Local::now() - last_location.time).num_minutes() > 5 {
-        None
-    } else {
-        Some(last_location.clone())
-    }
-}
-
 async fn inline_keyboard_button_pressed(
     bot: Bot,
-    dialoge: Dialogue<State, InMemStorage<State>>,
+    dialoge: DialogueState,
     cb: CallbackQuery,
     last_location: LastLocation,
     mode: Mode,
@@ -746,7 +197,7 @@ async fn inline_keyboard_button_pressed(
         Some("found:repeat") => {
             let mut frog = walk.frogs.last().unwrap().clone();
             frog.time = Local::now();
-            frog.gps_location = if_is_relevant(last_location);
+            frog.gps_location = last_location.as_location();
             walk.frogs.push(frog);
             is_repeat = true;
         }
@@ -895,7 +346,7 @@ async fn inline_keyboard_button_pressed(
 async fn maybe_end_walk(
     bot: Bot,
     walk: CompleteWalk,
-    dialoge: Dialogue<State, InMemStorage<State>>,
+    dialoge: DialogueState,
     mode: Mode,
     sent: SentMessage,
 ) -> Result<(), anyhow::Error> {
@@ -922,7 +373,7 @@ async fn maybe_end_walk(
 async fn end_walk(
     bot: Bot,
     mut walk: CompleteWalk,
-    dialoge: Dialogue<State, InMemStorage<State>>,
+    dialoge: DialogueState,
     mode: Mode,
 ) -> Result<(), anyhow::Error> {
     let date = Local::now();
@@ -987,51 +438,34 @@ impl GetChatId for UpdateWithSuppliedChatId {
     }
 }
 
-type LastLocation = Arc<Mutex<TimedLocation>>;
-
-#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize)]
-pub struct TimedLocation {
-    latitude: f64,
-    longitude: f64,
-    time: DateTime<Local>,
-}
-
-impl TimedLocation {
-    pub fn error() -> Self {
-        Self {
-            latitude: f64::NAN,
-            longitude: f64::NAN,
-            time: DateTime::UNIX_EPOCH.with_timezone(&Local),
-        }
-    }
-}
-
-async fn add_note(
+fn add_note(
     bot: Bot,
-    dialoge: Dialogue<State, InMemStorage<State>>,
+    dialoge: DialogueState,
     message: Message,
     text: String,
     location: LastLocation,
     sent: SentMessage,
-) -> anyhow::Result<()> {
-    let mut s = dialoge.get().await?.ok_or(anyhow!(
-        "This can only be executed if a walk has been started???"
-    ))?;
-    let classification = message
-        .reply_to_message()
-        .map(|m| sent.get(m.id))
-        .flatten()
-        .unwrap_or_default();
-    s.as_walk_mut()
-        .ok_or(anyhow!("There should be a walk at this point!"))?
-        .create_note(text, location, classification);
-    dialoge.update(s).await?;
-    bot.set_message_reaction(dialoge.chat_id(), message.id)
-        .reaction(vec![ReactionType::Emoji {
-            emoji: "✍".into()
-        }])
-        .await?;
-    Ok(())
+) -> impl Future<Output = anyhow::Result<()>> {
+    async move {
+        let mut s = dialoge.get().await?.ok_or(anyhow!(
+            "This can only be executed if a walk has been started???"
+        ))?;
+        let classification = message
+            .reply_to_message()
+            .map(|m| sent.get(m.id))
+            .flatten()
+            .unwrap_or_default();
+        s.as_walk_mut()
+            .ok_or(anyhow!("There should be a walk at this point!"))?
+            .create_note(text, location, classification);
+        dialoge.update(s).await?;
+        bot.set_message_reaction(dialoge.chat_id(), message.id)
+            .reaction(vec![ReactionType::Emoji {
+                emoji: "✍".into()
+            }])
+            .await?;
+        Ok(())
+    }
 }
 
 #[tokio::main]
@@ -1156,23 +590,7 @@ async fn main() -> anyhow::Result<()> {
                         .endpoint(State::frog_identified),
                 ),
         )
-        .branch(
-            Update::filter_edited_message()
-                .filter_map(|u: Update| {
-                    if let UpdateKind::EditedMessage(m) = u.kind {
-                        m.location().cloned()
-                    } else {
-                        None
-                    }
-                })
-                .endpoint(async |l: Location, tl: LastLocation| {
-                    let mut tl = tl.lock();
-                    tl.latitude = l.latitude;
-                    tl.longitude = l.longitude;
-                    tl.time = Local::now();
-                    Ok(())
-                }),
-        );
+        .branch(LastLocation::update_handler());
 
     Dispatcher::builder(bot, schema)
         .enable_ctrlc_handler()
